Add a simple PCT ELF image.
This PCT image prints out a few messages and then exits. 'make isoimage'
creates a CD image that passes this PCT ELF image to the kernel as the
initrd image. A future commit will remove this PCT binary and build a simple
'hello world' PCT ELF image as part of the kitten build.

Sample output:

[0]:<8>(TSK-0) hello world 3
[0]:<8>(TSK-0) this is a str
[0]:<8>(TSK-0) initialized number deadbeef
[0]:<8>(TSK-0) num=f00df00d
[0]:<8>(TSK-0) uninitialized number 0
[0]:<8>(TSK-0) argv[0]=pct
[0]:<8>(TSK-0) ENV1=(null)
[0]:<8>(TSK-0) ENV2=(null)
[0]:<8>(TSK-0) ENV3=(null)
[0]:<8>(TSK-0) rdtsc=1176882382
[0]:<8>(TSK-0) rdtsc=1181848571
[0]:<8>(TSK-0) rdtsc=1184371273