Run interrupt handlers with external interrupts disabled. This should be fine since interrupt handlers can't sleep or call schedule().
Also modify schedule() so that it remembers the external interrupt state
on entry and restores it on exit (it restores the state of the
scheduled task when it originally called schedule, which may have been
long ago and even on a different CPU). This enables schedule() to be
called with external interrupts disabled in entry.S and should prevent
interrupts from nesting in a single kernel stack.