Add kthread_create()
This patch adds a convienence function kthread_create() that does all
the under-the-cover work to create a kernel thread, setup the
parameters to it, clean up afterwards, etc.  It also adds an args[]
array to the struct start_state to make it easier to pass up to four
arguments to a new thread (either kernel or user level).