BUG_FIX: Use the correct lwIP netif input functions.
We were interfacing the device drivers using the approach for
non-concurrent OS environments, where everything ran in the same
thread context. Instead we should have used the interface for a
multi-threaded system. In the single threaded interface packets from a
device driver are operated on immediately as a result of ip_input. In
the multithreaded context, tcpip_input is used instead which just
implements a mailbox posting operation. The LWIP processing thread
then retrieves the packet from the mailbox.

From the webpage (http://www.sics.se/~adam/lwip/os.html):
"When run in a multi-threaded environment, incoming packets are
handled by the function tcpip_input(), which takes the same arguments
as the ip_input() function. The difference between the two functions
is that the tcpip_input() function does not process the incoming
packet immediately. It merely puts the packet on a queue which later
is drained by the TCP/IP thread."