BUG_FIX: Fix KFS infinite loop.
The issue was with kfs_mkdirent. If a file already exists in the hash table,
the code doesn't seem to care (despite a comment that seemed to imply that
it should). It overwrites the existing entry, and then reinserts it into
the hash table. This patch adds a flag to check if the entry was already there,
and if so does not do the hash table reinsertion. The state still gets
overwritten though.