BUG_FIX: Prevent bootmem bitmap from overlapping with kernel data.
This bug was inherited from Linux, fixed somewhere after 2.6.16.
Early on in boot the e820 map is consulted to find space
for the kernel page tables and bootmem allocator bitmap.  A byte
address is returned but the caller immediately converts it to a
page aligned address by shifting right by PAGE_SHIFT, destroying
the low order bits in the byte address.  This can result in the
allocated memory overlapping with kernel data, leading to
seemingly random behavior.  This was discovered while trying to
boot Kitten on a 24 GB system.  I'm guessing smaller memory systems
could also hit the bug, although I never saw it.