This patch adds the guest interruption interface to Kitten. This is the first step in multicore support.
When a guest is running on one core, but its vmm state is being
modified on another core, the changes will be invisible to the guest
until an exit is taken. For instance if a virtual device generates an
interrupt due to an event on another core, the virtual interrupt will
not be injected until the next exit. This interface allows a vmm
component to call for an explicit exit. Currently this is implemented
via the xcall_reschedule() API.

Palacios relies on Kitten to track the specific cpus that the guest is
running on. This takes the form of a mapping of logical guest cpus to
physical hardware cpus. So on an 8 core machine Kitten could assign
cores 3, 6, 7, and 8 to a 4 core guest, which Palacios would translate
to cpus 0, 1, 2, and 3. Kitten is in charge of keeping track of the
mapping, because it handles core scheduling. This allows Kitten to
migrate a VM between cores with hopefully minimal additional support.