Add support for signals and implement a proper task exit/cleanup path.
The exit path takes place in the context of the task that is exiting.
This is necessary because a task can be sleeping in the kernel and
be linked to various kernel lists/waitq's/etc.  The task therefore
has to be woken up so it can cleanup the kernel state that only
it knows about, and then call task_exit().  task_exit() doesn't
return.  The exited task's task structure is freed in schedule().

  - Restructure aspace and task subsystems.  Tasks now have aspace
    local IDs.  The aspace ID is the "process" ID and the task ID
    is the "thread" ID.
  - Implement signals, including sys_rt_sigprocmask, sig_rt_sigpending,
    sys_rt_sigaction, sys_kill, sys_wait4.  User-space can register
    signal handlers and have them fire appropriately.  When all
    tasks in an address space exit, the parent address space
    is sent a SIGCHLD.
  - Implement sys_exit() and sys_exit_group()
  - Modified hello_world init_task to wait for spawned threads to exit
    with pthread_join()
  - Move kthread_create() and friends to their own source file, kthread.c
  - Remove unhandled_signal, since it isn't used anywhere
  - Whitespace and other misc formatting changes.