Add sched_getcpu() implementation for RHEL5.
RHEL5 is ancient and does not provide a sched_getcpu() in its
glibc, even though the RHEL5 kernel has everything necessary. The
hello_world init task depends on sched_getcpu() so we provide our
own implementation. This fixes the build on RHEL5 and appears to
not break anything on newer distros with sched_getcpu() support.