BUGFIX: Put back zeroing of allocated pmem. Glibc assumes that sbrk() allocates zeroed memory, which depends on a process's heap being zeroed. Things fail in strange ways when the heap is not zeroed. Kitten was zeroing the heap originally, but at some point this was found to be a performance issues, causing IB timeouts at app launch.  If this is the case, the solution is to speed up the zeroing, rather than taking out the zeroing!  We'll have to deal with the speed issue the next time it is found to be an issue.... but the solution is not giving user-space uninitialized memory!
This was found when booting Kitten on the Teller system at Sandia.
The hello_world app would launch, but the first time printf() was
called to print a floating point number, a page fault would occur.
This was traced to glibc assuming that sbrk() returned zeroed memory.
This caused the thread local storage variable __libc_tsd_CTYPE_TOLOWER
pointer being uninitialized, which on Teller reliably resulted in it
being set to 0xFFFFFFFF.  It should have been 0.  An address calculation
used the 0xFFFFFFFF value, leading to a page fault.