Kitten userspace modifications. The PCT can now be configured to receive application information from an off node Linux job server. To enable this mode, use environment variable "WAIT_LAUNCH=1". Alternatively, the PCT can revert to its original format by not setting this variable.
Multi-node PMI support is added. PMI commits and barriers generated by
applications are sent to the server. Then, when an application issues a
get to its local key-val store, if the request misses, it is forwarded to
the server.

The runtime2/ codebase has been restructured to accommodate these changes.
The Linux job server is located in runtime2/pmi-server. A sample test
file is given in nidpid.txt.

Finally, the get.pl external script now generates 2 Portals source trees. It
configures 1 for Kitten and 1 for Linux.