Set task to state TASK_INTERRUPTIBLE before calling schedule() in timer_sleep_until()
This is needed because currently when a task calls timer_sleep_until(),and then it becomes
runnable soon, it skips the sleeping time. This happens for example when there is only one
task in the CPU runqueue. In that case, the task will run again inmediately after calling
schedule(), skipping the sleeping time. The same happens, for example, when Palacios vcores
call timer_sleep_until() through the v3_handle_halt function. Since the sleep-time is skipped,
the task keeps calling timer_sleep_until() too often, which is critical when using the
one-shot timer.