Add initial cut at in-kernel RCR driver.
The RCR driver starts up a kernel thread that periodically reads
a set of performance counters and publishes them to a 'blackboard'
that can be mapped into user space via mmap on /dev/rcr.

The set of performance counters monitored is hard-coded in
in the rcr_poller() kernel thread.  Currently, only two counters
are monitored -- socket 0 RAPL energy and the tsc timestamp counter.

An RCR test was added to the hello_world init task showing how
to open and map /dev/rcr and then read values from the blackboard.

This is the result of a two day hackfest (Feb 29 - Mar 1) with
Allan Porterfield, Stephen Olivier, and Kevin Pedretti