allow utime handling for more futex operations
The current implementation of futex() in kitten only does time handling
for FUTEX_WAIT, this is causing pthread_cond_timedwait to fail because
the utime value is not being filled out on futex_wake and being
preserved for subsequent waits. This may have uintended consequences so
I'm going to hold off on pushing it into the main tree until further
testing has been done.