Allow kitten xpmem to work with vader
This fixes some small problems that kept kitten xpmem from working with
vader. First it does not require null in the XPMEM_PERMIT_MODE so we
disable that check. Second the smartmap to xpmem conversion made
assumptions that don't work with our current MPI setup so I kludged it
so that smartmaps and xpmem regions match each other.